The LAGO ANNA v1 package was restructured following standards. Major changes included:
* New header at each file.
* New license description file.
* Header files are now in the include/ directory.
* App source files are now in the src/ directory.
* Makefile was adapted for the new tree structure.
* Versioning was eliminated from each file to avoid future issues.
* The README was completely rewriten following the schema used for LAGO ARTI and LAGO ACQUA.
* App integrated help was improved
