+++ include/lago_defs.h
@@ -1,0 +1,31 @@
+/*
+################################################################################
+# File:        lago_data.h
+# Description: Global definitions for the LAGO ANNA framework
+# Author:      Hernán Asorey
+# Email:       asoreyh@gmail.com
+# Date:        2012
+# 
+# Copyright:   [2012] [The LAGO Collaboration]
+# License:     BSD-3-Clause
+# See the LICENSE file in the project root for full license information.
+################################################################################
+*/
+#ifndef LAGO_DEFS_H
+#define LAGO_DEFS_H 
+
+#define PROJECT "The LAGO ANNA Suite"
+#define CODENAME "ANNA"
+// DATAVERSION AND CODEVERSION SHOULD BE THE SAME NUMBER
+#define CODEVERSION "v5"
+#define DATAVERSION 5 
+
+#define CHANNELS 3
+#define TRACELEN 12
+#define BASELINE 50
+#define BIN 25.
+#define ADCMAX 1024
+#define CHRGMAX 4096
+#define TRIGGERBIN 3
+
+#endif
